Prevent 3 astronomy images from loading in quick succession
This extension has an issue that it quickly loads ~3 astronomy images in quick succession, because the `onUpdateRequest` method automatically fires ~3 times. By changing this to the `onActivateRequest` method, it fixes the issue and loads just one image, while clicking the "Random Astronomy Picture" command in the command pallet still loads a new image.